requirements.txt
requests==2.31.0
python-dotenv==1.0.0
yfinance==0.2.37
pandas==2.2.1
numpy==1.26.4
ta==0.11.0
flask==3.0.2
flask-cors==4.0.0
